+++ CHANGELOG.md
@@ -17,6 +17,10 @@ As this project is a user-facing application, the places in the semantic version
 
 - Sort sources files by name
 
+### Changed
+
+- Removed storage of resource data in browser
+
 ## [1.12.2](https://github.com/spraakbanken/mink-frontend/releases/tag/v1.12.2) (2025-09-26)
 
 ### Fixed